feat(memory): user-editable memory REST API with per-scope authorization (Phase 2)
六端点开放记忆模块给用户自管理：GET/POST /memory、GET/PUT/DELETE
/memory/{id}、POST /memory/recall（召回预览，带维度标签不改状态）。
授权矩阵按维度：organization=org owner/admin；project=owner/maintainer；
agent=agent owner 或 org 管理员；user=仅本人；session=系统托管拒绝手写。
人工创建/编辑记忆打 human_edited 信任信号（迁移 000027），召回排序加权
（confidence +10 加成）；store 补 list/get/update/forget_by_id（软删）。
修复迁移 000026 MySQL 隐患：source_task_id INT→BIGINT（tasks.id 为
BIGINT，FK 类型不兼容会让 MySQL 部署在建表时直接失败，SQLite 测不出）。
+16 用例（授权矩阵/租户边界/去重/编辑/软删/跨组织 404/召回预览/
human_edited 召回偏好），全量门禁 2310 passed。

+++ migrations/add_agent_memory.py
@@ -51,7 +51,7 @@ def upgrade(connection):
                 title VARCHAR(500) NOT NULL,
                 content TEXT NOT NULL,
                 source_type VARCHAR(50) NULL,
-                source_task_id INT NULL,
+                source_task_id BIGINT NULL,
                 agent_id INT NULL,
                 confidence INT NULL,
                 is_valid INT NOT NULL DEFAULT 1,
